+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * Administration page
 * Icon uploading and management
 * Uploading of multiple smiles at the same time
-* Suggestions: user can offer to add a smile to the collection (in development)
+* Suggestions: user can offer to add a smile to the collection
 * Editing and forking of smilepacks
 * The ability to create multiple categories with the same name (including empty name) in smilepacks
 * Many bug fixes (including frontend fixes for Safari and IE11)